chore: Move rhum into tests deps to keep out of main dep tree
