+++ deps.ts
@@ -1,3 +1,2 @@
 export * as colors from "https://deno.land/std@0.97.0/fmt/colors.ts";
 export { serve } from "https://deno.land/std@0.97.0/http/server.ts";
-export { Rhum } from "https://deno.land/x/rhum@v1.1.10/mod.ts";
