+++ tests/deps.ts
@@ -1,0 +1,1 @@
+export { Rhum } from "https://deno.land/x/rhum@v1.1.10/mod.ts";
